refactoring of json serialisation handling
